Revert "implemented apply job feature "
