+++ utils/jobs.json
@@ -1,1 +1,1 @@
-[{"name":"john","age":"21","education":"diploma in information technology","phone":"99998888"},{"name":"sam","age":"21","education":"diploma in information technology","phone":"88887777"}]
+[]
